Removed notice suppression in favor of checking to see if running in the CLI environment.
The notice suppression had been added in r21712.

When phpCAS is included in a command-line application (such as the unit tests) none of
the REQUEST_URI, SCRIPT_NAME, or QUERY_STRING elements of the $_SERVER array are set.
While phpCAS is unlikely to operate in a meaningful way from the CLI, it should be able
to at least be included along with other libraries without throwing notices.

This commit also includes change from r21556: PHPCAS-81 fix wrong check of the REQUEST_URI
for IIS that throws an unwanted notice


git-svn-id: https://source.jasig.org/cas-clients/phpcas/branches/PHPCAS-54@21788 f5dbab47-78f9-eb45-b975-e544023573eb