PHPCAS-90 Added a RequestInterface::getResponseStatusCode() method.
In testing the ProxiedService system I noticed that the RequestInterface::sent()
only returns false on network or system errors, not on requests that have a
valid HTTP response that is an error response.

To allow clients an easier way to validate that the response was OK, I've added
a RequestInterface::getResponseStatusCode() method to prevent clients from having
to parse the response headers themselves just to check the status.

git-svn-id: https://source.jasig.org/cas-clients/phpcas/branches/PHPCAS-90@23506 f5dbab47-78f9-eb45-b975-e544023573eb