PHPCAS-90 Fixed CurlRequest cookie setting so that it properly sets array(name => value) cookies.
The documented behavior of CAS_RequestInterface::addCookies(array $cookies) is that
the cookie array is of the form
     array('cookie_name' => 'cookie_value', 'cookie_name2' => cookie_value2')

However, CAS_CurlRequest::_sendRequest() was assuming that the cookie array was in the form
     array('cookie_name=cookie_value', 'cookiename2=cookievalue2')
This array format was what CASClient::readURL() was passing, but broke our new
ProxiedService classes that were adhering to the interface description.

Since I have removed cookie support from CASClient::readURL() in the last commit, there
is now no conflict and this implementation can be updated to match the interface description.

git-svn-id: https://source.jasig.org/cas-clients/phpcas/branches/PHPCAS-90@23512 f5dbab47-78f9-eb45-b975-e544023573eb