Cookies with Domain=.hostname.domain.edu are now valid for hostname.domain.edu
We were being a bit too strict in our hostname matching by now allowing cookies
to be valid for a host when a '.' was prepended.

According to the PHP documentation (http://php.net/manual/en/function.setcookie.php)
RFC 2109 should be interpreted as requiring that in addition to
Cookies with Domain=.hostname.domain.edu being valid for subdomains of
hostname.domain.edu the cookie is also valid for hostname.domain.edu itself.
This is the behavior browsers implement as well.