#16 Fix for case where already authenticated, but given new proxy ticket.
When a proxied service receives a second request with a valid session cookie
the authentication is still valid. If a proxy ticket is also included it
is superfluous and can be ignored while using the existing authenticated
session.

This change ensures that isAuthenticated() continues to return true based on
the valid session cookie, ignoring the extra proxy ticket.