#203 - Fix errors in source/CAS.php when run under phpdbg
This code assumed any SAPI other than "cli" is a web request, which is not the
case for phpdbg, the debugger that ships with PHP 5.6. Since phpdbg is
implemented as a SAPI module, the php_sapi_name() function will return "phpdbg".
If you haven't set up the $_SERVER superglobal beforehand, this code would
trigger "Undefined index" errors when run under phpdbg.

I opted to fix this by removing the php_sapi_name() check, as it's not
necessary. Just checking the superglobals with isset() is sufficient.