Add support for logging via a PSR-3 logger (#329)
PSR-3 is a PHP standard for cross-library inter-operation and
compatibility. Details on the specifics of PSR-3 are at:

https://www.php-fig.org/psr/psr-3/

Uses the psr/log package to enforce the PSR-3 LoggerInterface.

https://packagist.org/packages/psr/log

The old phpCAS::setDebug() is deprecated in favor using a logger, but
will continue to work for backward compatibility. Users that wish to
continue logging to a flat file with the new interface can do so using a
PSR-3 logger such as Monolog's StreamHandler.