Fix output during tests (#336)
For tests that write to the output buffer, capture and clean the output
buffer. Handle both the success case and the failure case.
Unfortunately, the finally keyword was not introduced until PHP 5.5.

Remove the unused @outputBuffering decorator in comments.

Remove the unreachable code `$this->assertFalse($result);` after the
exception is expected to be thrown.