Fix PHPUnit GitHub actions (#388)
* Specify composer as the phpunit bootstrap

The PHPUnit action says the [default bootstrap is
vendor/autoload.php](https://github.com/marketplace/actions/phpunit-php-actions#inputs)
but it seems that the composer autoload was not called.

* Autoload CAS.php for tests

The CAS.php file defines a number of constants used by tests. Because
tests call directly into the underlying class and do not pass through
the CAS.php entrypoint, these constants are not defined in tests.